Make sync stream helper a common util
