+++ plugin-flex-ts-template-v2/src/feature-library/admin-ui/custom-components/AdminView/AdminView.tsx
@@ -22,7 +22,7 @@ import { getFeatureFlagsUser } from '../../../../utils/configuration';
 import FeatureCard from '../FeatureCard';
 import AdminUiService from '../../utils/AdminUiService';
 import { saveUserConfig, saveGlobalConfig, shouldShowFeature, featureCommon } from '../../utils/helpers';
-import { subscribe, unsubscribe, publishMessage, SyncStreamEvent } from '../../utils/sync-stream';
+import { subscribe, unsubscribe, publishMessage, SyncStreamEvent } from '../../../../utils/sdk-clients/sync/SyncClient';
 import { AdminUiNotification } from '../../flex-hooks/notifications';
 import FeatureModal from '../FeatureModal';
 
@@ -37,15 +37,17 @@ const AdminView = () => {
   const [config, setConfig] = useState({} as any);
   const [globalConfig, setGlobalConfig] = useState({} as any);
   const [userConfig, setUserConfig] = useState({} as any);
+  const [syncStream, setSyncStream] = useState(null as any); // twilio-sync does not export the SyncStream type
 
   const strings = Manager.getInstance().strings as any;
+  const streamName = 'template-admin';
   const streamEvent = 'template-admin-update';
 
   useEffect(() => {
     initialize();
     return () => {
       Notifications.dismissNotificationById(AdminUiNotification.SAVE_DISABLED);
-      unsubscribe();
+      unsubscribe(syncStream);
     };
   }, []);
 
@@ -73,7 +75,7 @@ const AdminView = () => {
     setIsUpdatedByOtherUser(false);
     setIsUpdatedModalOpen(false);
     setIsLoading(true);
-    await subscribe(handleSyncMessage);
+    setSyncStream(await subscribe(streamName, handleSyncMessage));
     await reloadGlobalConfig();
     reloadUserConfig();
     setIsLoading(false);
@@ -110,7 +112,7 @@ const AdminView = () => {
       const saveResult = await saveGlobalConfig(feature, config, mergeFeature);
 
       if (saveResult) {
-        publishMessage({ event: streamEvent });
+        publishMessage(syncStream, { event: streamEvent });
         setGlobalConfig(saveResult);
         return true;
       }
